Fix tests and multi file mode
