+++ Source/AutoExtern.Test/Minimal/Library1.cs
@@ -1,0 +1,5 @@
+namespace NS;
+
+public interface Intf {
+  public string Prop { get; }
+}
