+++ Source/AutoExtern.Test/Minimal/Library2.cs
@@ -1,9 +1,5 @@
 namespace NS;
 
-public interface Intf {
-  public string Prop { get; }
-}
-
 public class Impl : Intf {
   public int Field;
   public string Prop => Field.ToString();
